fix(api): split graph-payload-loaders under file-length gate
Move bulk full-graph and repo loaders to graph-payload-bulk.ts so
staging CI passes the 300-line cap again.
